IaaS/Development/Backup_and_Recovery/requirements.txt
boto3>=1.26.0
python-libvirt>=8.0.0
PyYAML>=6.0
schedule>=1.1.0
cryptography>=39.0.0
paramiko>=3.0.0
click>=8.0.0
python-dateutil>=2.8.2
requests>=2.28.0
tenacity>=8.0.1
